requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
SQLAlchemy==2.0.27
psycopg-binary==3.2.12
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
pywebpush==2.1.2
python-dotenv==1.0.0
boto3==1.26.163
alembic==1.12.0
py_vapid==1.9.2
httpx==0.24.1
cryptography==41.0.3
